Move closure/generator type info methods to TyCtxt
